refactor(dashboard): split OverviewBento into sectioned components with per-section error boundaries
Splits the 2,261-line OverviewBento.tsx into 15 focused sub-components
under dashboard/src/components/overview/, each wrapped in a new
SectionErrorBoundary so a single failing API response no longer crashes
the whole page. Fetch logic is extracted into a useOverviewData hook.
The page-level orchestrator is now 241 lines and preserves all existing
behavior, styling, aria attributes, and the LayoutSwitcher export
consumed by OverviewFocus.
